requirements.txt
Flask==2.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==1.0.0
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-api-python-client==2.32.0
google-cloud-secret-manager==2.12.0
pypdf2==2.10.5
python-dotenv==0.19.2
psycopg2-binary==2.9.3
gunicorn==20.1.0
Werkzeug==2.0.1
